requirements.txt
PyPDF2==3.0.1
python-docx==1.1.2